OpenTelemetry-Jaeger/requirements.txt
fastapi
uvicorn
pydantic
opentelemetry-distro
opentelemetry-exporter-otlp-proto-grpc
opentelemetry-instrumentation-fastapi